finish class number selection function
