+++ project/main.py
@@ -23,9 +23,6 @@ from typing import Any
 from pytorch_lightning import Trainer, seed_everything
 from pytorch_lightning.loggers import TensorBoardLogger
 
-from imblearn.over_sampling import RandomOverSampler
-from imblearn.under_sampling import RandomUnderSampler
-
 # callbacks
 from pytorch_lightning.callbacks import (
     TQDMProgressBar,
@@ -39,10 +36,7 @@ from dataloader.data_loader import WalkDataModule
 from train import GaitCycleLightningModule
 
 import hydra
-
-from sklearn.model_selection import StratifiedGroupKFold, train_test_split, GroupKFold
-from pathlib import Path
-
+from cross_validation import DefineCrossValidation
 
 def train(hparams, dataset_idx, fold):
     seed_everything(42, workers=True)
@@ -103,206 +97,6 @@ def train(hparams, dataset_idx, fold):
     # the validate method will wirte in the same log twice, so use the test method.
     trainer.test(classification_module, data_module, ckpt_path="best")
 
-
-class DefineCrossValidation:
-    """process:
-    cross validation > over/under sampler > train/val split
-    fold: [train/val]: [path]
-    """
-
-    def __init__(self, config) -> None:
-        
-        self.video_path = Path(config.data.gait_seg_data_path) # json file path
-        self.gait_seg_idx_path = Path(config.data.gait_seg_index_data_path) # used for training path mapping
-        self.K = config.train.fold
-        self.sampler = config.data.sampling # data balance, [over, under, none]
-
-        self.class_num = config.model.model.class_num
-
-    @staticmethod
-    def random_sampler(X: list, y: list, train_idx: list, val_idx: list, sampler):
-        # train
-        train_mapped_path = []
-        new_X_path = [X[i] for i in train_idx]
-
-        sampled_X, sampled_y = sampler.fit_resample(
-            [[i] for i in range(len(new_X_path))], [y[i] for i in train_idx]
-        )
-
-        # map sampled_X to new_X_path
-        for i in sampled_X:
-            train_mapped_path.append(new_X_path[i[0]])
-
-        # val
-        val_mapped_path = []
-        new_X_path = [X[i] for i in val_idx]
-
-        sampled_X, sampled_y = sampler.fit_resample(
-            [[i] for i in range(len(new_X_path))], [y[i] for i in val_idx]
-        )
-
-        # map
-        for i in sampled_X:
-            val_mapped_path.append(new_X_path[i[0]])
-
-        return train_mapped_path, val_mapped_path
-
-    @staticmethod
-    def process_cross_validation(video_path):
-        _path = video_path
-
-        X = []  # patient index
-        y = []  # patient class index
-        groups = []  # different patient groups
-        # process one disease in one loop.
-        for disease in _path.iterdir():
-            if disease.name != "log":
-                patient_list = sorted(list(disease.iterdir()))
-                name_map = set()
-
-                for p in patient_list:
-                    name, _ = p.name.split("-")
-                    name_map.add(name)
-
-                element_to_num = {element: idx for idx, element in enumerate(name_map)}
-
-                for i in range(len(patient_list)):
-                    name, _ = patient_list[i].name.split("-")
-                    # load the video tensor from json file
-                    with open(patient_list[i]) as f:
-                        file_info_dict = json.load(f)
-
-                    label = file_info_dict["label"]
-
-                    X.append(patient_list[i])  # true path in Path
-                    y.append(label)  # label, 0, 1, 2
-                    groups.append(element_to_num[name])  # number of different patient
-
-        return X, y, groups
-
-    def make_dataset_with_video(self, val_dataset_idx: list, fold: int, flag: str):
-        temp_path = self.gait_seg_idx_path / self.sampler / str(fold) / str(flag)
-        val_idx = val_dataset_idx
-
-        shutil.rmtree(temp_path, ignore_errors=True)
-
-        for path in val_idx:
-            with open(path) as f:
-                file_info_dict = json.load(f)
-
-            video_name = file_info_dict["video_name"]
-            video_path = file_info_dict["video_path"]
-            video_disease = file_info_dict["disease"]
-
-            if not (temp_path / video_disease).exists():
-                (temp_path / video_disease).mkdir(parents=True, exist_ok=False)
-
-            shutil.copy(
-                video_path, temp_path / video_disease / (video_name + ".mp4")
-            )
-
-        return temp_path
-
-    def prepare(self):
-        """define cross validation first, with the K.
-        #! the 1 fold and K fold should return the same format.
-        fold: [train/val]: [path]
-
-        Args:
-            video_path (str): the index of the video path, in .json format.
-            K (int, optional): crossed number of validation. Defaults to 5, can be 1 or K.
-
-        Returns:
-            list: the format like upper.
-        """
-        K = self.K
-
-        ans_fold = {}
-
-        # define the cross validation
-        # X: video path, in path.Path foramt. len = 1954
-        # y: label, in list format. len = 1954, three type > [0, 1, 2]
-        # groups: different patient, in list format. It means unique patient index. [54]
-        X, y, groups = self.process_cross_validation(self.video_path)
-
-        sgkf = StratifiedGroupKFold(n_splits=K)
-
-        for i, (train_index, test_index) in enumerate(
-            sgkf.split(X=X, y=y, groups=groups)
-        ):
-            if self.sampler in ["over", "under"]:
-                if self.sampler == "over":
-                    ros = RandomOverSampler(random_state=42)
-                elif self.sampler == "under":
-                    ros = RandomUnderSampler(random_state=42)
-
-                train_mapped_path, val_mapped_path = self.random_sampler(
-                    X, y, train_index, test_index, ros
-                )
-
-            else:
-                train_mapped_path = [X[i] for i in train_index]
-                val_mapped_path = [X[i] for i in test_index]
-
-            # make the val data path
-            train_video_path = self.make_dataset_with_video(train_mapped_path, i, "train")
-            val_video_path = self.make_dataset_with_video(val_mapped_path, i, "val")
-
-            ans_fold[i] = [train_mapped_path, train_video_path, val_video_path]
-
-        return ans_fold, X, y, groups
-
-    def __call__(self, *args: Any, **kwds: Any) -> Any:
-
-        # when json file changed, need to reprocess the dataset.
-        if not os.path.exists(self.gait_seg_idx_path/ self.sampler):
-            fold_dataset_idx, *_ = self.prepare()
-
-            json_fold_dataset_idx = copy.deepcopy(fold_dataset_idx)
-
-            for k, v in fold_dataset_idx.items():
-                
-                # train mapping path, include the gait cycle index
-                train_mapping_idx = v[0]
-
-                json_fold_dataset_idx[k][0] = [str(i) for i in train_mapping_idx]
-
-                # train video path
-                train_video_idx = v[1]
-                json_fold_dataset_idx[k][1] = str(train_video_idx)
-
-                # val video path
-                val_dataset_idx = v[2]
-                json_fold_dataset_idx[k][2] = str(val_dataset_idx)
-
-            with open((self.gait_seg_idx_path / self.sampler / "index.json"), "w") as f:
-                json.dump(json_fold_dataset_idx, f)
-
-
-        elif os.path.exists(self.gait_seg_idx_path):
-            with open(self.gait_seg_idx_path / self.sampler / "index.json", "r") as f:
-                fold_dataset_idx = json.load(f)
-
-            # unpack the 
-            for k, v in fold_dataset_idx.items():
-                # train mapping, include the gait cycle index
-                train_mapping_idx = v[0]
-                fold_dataset_idx[k][0] = [Path(i) for i in train_mapping_idx]
-
-                # train video path
-                train_video_idx = v[1]
-                fold_dataset_idx[k][1] = Path(train_video_idx)
-
-                # val
-                val_dataset_idx = v[2]
-                fold_dataset_idx[k][2] = Path(val_dataset_idx)
-                
-        else:
-            raise ValueError(
-                "the gait seg idx path is not exist, please check the path."
-            )
-
-        return fold_dataset_idx
 
 
 @hydra.main(
