+++ project/train.py
@@ -17,6 +17,8 @@ HISTORY:
 Date 	By 	Comments
 ------------------------------------------------
 
+22-03-2024	Kaixu Chen	add different class number mapping, now the class number is a hyperparameter.
+
 14-12-2023	Kaixu Chen refactor the code, now it a simple code to train video frame from dataloader.
 
 '''
